refactor: renamed footer component
