+++ packages/gatsby-theme-newrelic/src/components/GlobalFooter.js
